# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant Chain POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /outlets:
    get:
      operationId: listOutlets
      tags: [outlets]
      summary: List all outlets
      responses:
        "200":
          description: List of outlets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Outlet"
    post:
      operationId: createOutlet
      tags: [outlets]
      summary: Create an outlet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutletInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outlet"
  /outlets/{id}:
    get:
      operationId: getOutlet
      tags: [outlets]
      summary: Get outlet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Outlet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outlet"
        "404":
          description: Not found
    patch:
      operationId: updateOutlet
      tags: [outlets]
      summary: Update an outlet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutletUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Outlet"
    delete:
      operationId: deleteOutlet
      tags: [outlets]
      summary: Delete an outlet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Staff ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Outlet:
      type: object
      required: [id, name, address, phone, taxRate, currency, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        taxRate:
          type: number
        currency:
          type: string
        createdAt:
          type: string
    OutletInput:
      type: object
      required: [name, address, phone, taxRate, currency]
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        taxRate:
          type: number
        currency:
          type: string
    OutletUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        taxRate:
          type: number
        currency:
          type: string

    # ── Staff schemas ─────────────────────────────────────────────────────
...